fix(docs): keep process.env out of the client bundle; stop 404 advertising a .md
Split the browser-safe markdownSiblingForPathname into src/llmsPaths.js so the
client bundle no longer imports llmsTxt.js, whose top-level process.env.ENV threw
"process is not defined" on every page during hydration (Docusaurus/webpack do not
define arbitrary process.env.* in the browser). llmsTxt.js keeps the build-only
plugin/remark wiring and is now required only by docusaurus.config.ts.

Also add 404.html to EXCLUDED_SEGMENTS so the rendered 404 page stops advertising a
non-existent /docs/404.html.md markdown sibling (its pathname is /404.html, whose
first segment was not previously excluded).
